+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Allow icon buttons to have an ARIA label
 - Set a 'Clear' label on the SearchBox clear icon button
+- Allow Level component to have an ARIA label
 
 ## 0.0.42